fix: post-deploy payment, timezone and UI bugs; add feature guide
Payments
- Add trip "pay now" and wallet top-up screens; neither existed, so any
  non-cash trip stayed unpaid forever and drivers were never credited.
- Route gateway returns through POST/GET /payments/:id/return, which
  verifies with SSLCommerz, settles immediately, and 303-redirects to a
  new client result page (SSLCommerz posts back, which a static host
  cannot serve; the old client URL was also a 404).
- Mark cancelled/failed gateway attempts failed, and stop attempts
  older than 30 minutes from blocking a retry.

Time zones
- Run every DB session in Asia/Dhaka so daily earnings, views and
  stats use Bangladesh days (Supabase defaults to UTC, filing
  00:00-06:00 trips under the previous day).
- Use Dhaka dates for server defaults and client date inputs; the
  earnings screen hid overnight trips.

Other
- CLIENT_ORIGIN accepts a comma-separated list for CORS.
- Keyboard Enter/Space confirms the driver trip slider.
- Map route summary no longer overlaps SOS / booking hint on phones.
- Earnings trip counts returned as numbers, not strings.
- Stop logging successful /health probes.

Verified: full ride end to end in production mode (book, match,
arrive, start, complete, pay through SSLCommerz sandbox, rate), a
33-screen crawl across all roles, and 351 server tests.

+++ server/src/repositories/earnings.repository.js
@@ -17,7 +17,7 @@ export async function insertEarning(
 
 export async function listDailyForDriver(driverId, { from, to }, client = pool) {
   const { rows } = await client.query(
-    `SELECT earning_date::text AS "earningDate", trips_count AS "tripsCount",
+    `SELECT earning_date::text AS "earningDate", trips_count::int AS "tripsCount",
             gross_total AS "grossTotal", commission_total AS "commissionTotal", net_total AS "netTotal"
      FROM v_driver_daily_earnings
      WHERE driver_id = $1 AND earning_date BETWEEN $2 AND $3
